+++ Shaders/TerrainDetailLit.shader
@@ -26,8 +26,8 @@
 			{
 				float4 positionOS : POSITION;
 				float2 texcoord : TEXCOORD0;
-				float3 normalOS : NORMAL;
-				float4 color : COLOR;
+				half3 normalOS : NORMAL;
+				half4 color : COLOR;
 			};
 
 			struct Varyings
